Protecting resources with autoriztion

+++ backend/routes/posts.js
@@ -71,10 +71,14 @@ checkAuth,
     content: req.body.content,
     imagePath: imagePath
   })
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then(result => {
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if(result.nModified > 0){
+        res.status(200).json({message: 'update successful'});
+      } else {
+        res.status(401).json({message: 'Not authorized'});
+      }
 
-    res.status(200).json({message: 'update successful'});
+
   });
 });
 
@@ -134,8 +138,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result =>{
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result =>{
     console.log(result);
+    if(result.n > 0){
+      res.status(200).json({message: 'Deletion successful'});
+    } else {
+      res.status(401).json({message: 'Not authorized'});
+    }
   res.status(200).json({message: "post deleted"});
 });
 });
